+++ scout/commands/load/case.py
@@ -15,7 +15,7 @@ from scout.exceptions import IntegrityError, ConfigError
 LOG = logging.getLogger(__name__)
 
 @click.command('case', short_help='Load a case')
-@click.option('--vcf', 
+@click.option('--vcf',
     type=click.Path(exists=True),
     help='path to clinical VCF file to be loaded'
 )
@@ -23,39 +23,39 @@ LOG = logging.getLogger(__name__)
     type=click.Path(exists=True),
     help='path to clinical SV VCF file to be loaded'
 )
-@click.option('--vcf-cancer', 
+@click.option('--vcf-cancer',
     type=click.Path(exists=True),
     help='path to clinical cancer VCF file to be loaded'
 )
-@click.option('--vcf-str', 
+@click.option('--vcf-str',
     type=click.Path(exists=True),
     help='path to clinical STR VCF file to be loaded'
-)             
-@click.option('--owner', 
-    help='parent institute for the case', 
+)
+@click.option('--owner',
+    help='parent institute for the case',
     default='test'
 )
 @click.option('--ped',
     type=click.File('r')
 )
-@click.option('-u', '--update', 
+@click.option('-u', '--update',
     is_flag=True
 )
-@click.option('--no-variants', 
+@click.option('--no-variants',
     is_flag=False
 )
-@click.argument('config', 
+@click.argument('config',
     type=click.File('r'), required=False
 )
-@click.option('--peddy-ped', 
+@click.option('--peddy-ped',
     type=click.Path(exists=True),
     help='path to a peddy.ped file'
 )
-@click.option('--peddy-sex', 
+@click.option('--peddy-sex',
     type=click.Path(exists=True),
     help='path to a sex_check.csv file'
 )
-@click.option('--peddy-check', 
+@click.option('--peddy-check',
     type=click.Path(exists=True),
     help='path to a ped_check.csv file'
 )
@@ -74,8 +74,8 @@ def case(context, vcf, vcf_sv, vcf_cancer, vcf_str, owner, ped, update, config,
 
     # Scout needs a config object with the neccessary information
     # If no config is used create a dictionary
-    config_raw = yaml.load(config) if config else {}
-    
+    config_raw = yaml.load(config, Loader=yaml.FullLoader) if config else {}
+
     try:
         config_data = parse_case_data(
             config=config_raw,
@@ -94,7 +94,7 @@ def case(context, vcf, vcf_sv, vcf_cancer, vcf_str, owner, ped, update, config,
         context.abort()
 
     LOG.info("Use family %s" % config_data['family'])
-    
+
     try:
         case_obj = adapter.load_case(config_data, update)
     except Exception as err: