+++ CHANGELOG.md
@@ -4,11 +4,24 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
+
 ## [4.x.x]
 
 ### Added
+- Search for similar cases from scout view CLI
+
 
 ### Fixed
+
+
+
+## [4.4.1]
+
+### Added
+- Display SV rank model version when available
+
+### Fixed
+- Fixed upload of delivery report via API
 
 
 ## [4.4.0]